+++ client/src/components/schedule/TripInfo.tsx
